backend/pyproject.toml
[project]
name = "tito-ai-backend"
version = "0.1.0"
description = "Tito AI Backend - Voice AI Assistant"
authors = [{name = "Tito AI Team"}]
readme = "README.md"
requires-python = ">=3.11,<3.13"
dependencies = [
    "aiohttp>=3.11.12",
    "python-dotenv==1.0.1",
    "fastapi==0.115.8",
    "uvicorn==0.34.0",
    "loguru==0.7.3",
    "pytz==2025.1",
    "pipecat-ai[daily,openai,deepgram,cartesia,silero,google,anthropic,groq,together,mistral,playht,azure,gladia,assemblyai,aws]",
    "pipecat-ai-flows",
    "google-genai",
    "openai",
    "anthropic",
    "groq",
    "together",
    "mistralai",
    "ultravox-client",
    "panoramisk>=1.4",
]

[project.optional-dependencies]
dev = [
    "pytest",
    "pytest-asyncio",
    "black",
    "isort",
    "flake8",
    "mypy",
]

[dependency-groups]
dev = [
    "pytest>=7.0.0",
    "pytest-asyncio>=0.21.0",
    "black>=23.0.0",
    "isort>=5.12.0",
    "flake8>=6.0.0",
    "mypy>=1.0.0",
]

[tool.black]
line-length = 100
target-version = ['py310']

[tool.isort]
profile = "black"
line_length = 100

[tool.mypy]
python_version = "3.10"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true

[tool.hatch.build.targets.wheel]
packages = ["app"]
